backend/requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0

# API Clients
tweepy==4.14.0
openai==1.3.0
twilio==8.10.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
bcrypt==4.1.1

# ML/AI
transformers==4.35.0
torch==2.1.0
sentencepiece==0.1.99

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.1

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.0

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0

